[FIX] mail: scroll to not yet loaded pinned message
Before this PR, clicking on the "user pinned a message to the channel"
notification linked to a not yet loaded message would not scroll to
the pinned message.

This issue arises because the `highlightMessage` function does nothing
if the message's thread is different from the current one. However,
when a message is unknown, the thread is also unknown, and the
function does nothing.

The notification is only displayed in the origin thread of the
message, so we can safely pass this information, assuming the message
thread and the notification thread are the same.
