+++ addons/mail/static/src/core/common/thread.js
@@ -509,6 +509,7 @@ export class Thread extends Component {
                     id: Number(oeId),
                     res_id: this.props.thread.id,
                     model: this.props.thread.model,
+                    thread: this.props.thread,
                 }),
                 this.props.thread
             );
