+++ addons/mail/static/tests/thread/unread_messages_banner.test.js
@@ -10,17 +10,9 @@ import {
     startServer,
     step,
 } from "@mail/../tests/mail_test_helpers";
-import { Thread } from "@mail/core/common/thread";
 import { describe, test } from "@odoo/hoot";
 import { tick } from "@odoo/hoot-mock";
-import {
-    Command,
-    getService,
-    onRpc,
-    patchWithCleanup,
-    serverState,
-    withUser,
-} from "@web/../tests/web_test_helpers";
+import { Command, getService, onRpc, serverState, withUser } from "@web/../tests/web_test_helpers";
 import { rpc } from "@web/core/network/rpc";
 
 describe.current.tags("desktop");
@@ -66,49 +58,6 @@ test("mark thread as read from unread messages banner", async () => {
         parent: ["span", { text: "30 new messagesMark as Read" }],
     });
     await contains(".o-mail-Thread-jumpToUnread", { count: 0 });
-});
-
-test("scroll to the first unread message (slow ref registration)", async () => {
-    const pyEnv = await startServer();
-    const channelId = pyEnv["discuss.channel"].create({ name: "general" });
-    const messageIds = [];
-    for (let i = 0; i < 201; i++) {
-        messageIds.push(
-            pyEnv["mail.message"].create({
-                body: `message ${i}`,
-                model: "discuss.channel",
-                res_id: channelId,
-            })
-        );
-    }
-    const [selfMember] = pyEnv["discuss.channel.member"].search([
-        ["partner_id", "=", serverState.partnerId],
-        ["channel_id", "=", channelId],
-    ]);
-    pyEnv["discuss.channel.member"].write([selfMember], { new_message_separator: messageIds[100] });
-    let slowRegisterMessageRef = false;
-    patchWithCleanup(Thread.prototype, {
-        async registerMessageRef() {
-            if (slowRegisterMessageRef) {
-                // Ensure scroll is made even when messages are mounted later.
-                await new Promise((res) => setTimeout(res, 250));
-            }
-            super.registerMessageRef(...arguments);
-        },
-    });
-    await start();
-    await openDiscuss(channelId);
-    await tick(); // wait for the scroll to first unread to complete
-    await click(".o-mail-Thread-banner", {
-        text: "You're viewing older messagesJump to Present",
-    });
-    await isInViewportOf(".o-mail-Message:contains(message 200)", ".o-mail-Thread");
-    slowRegisterMessageRef = true;
-    await click("span", {
-        text: "101 new messages",
-        parent: ["span", { text: "101 new messagesMark as Read" }],
-    });
-    await isInViewportOf(".o-mail-Message:contains(message 100)", ".o-mail-Thread");
 });
 
 test("scroll to unread notification", async () => {
